+++ src/Matchers/ModifiedMatcher.php
@@ -2,6 +2,7 @@
 
 namespace SmartonDev\HttpCache\Matchers;
 
+use SmartonDev\HttpCache\Exceptions\DateMalformedStringException;
 use SmartonDev\HttpCache\Helpers\HttpHeaderHelper;
 use SmartonDev\HttpCache\Helpers\TimeHelper;
 
@@ -13,7 +14,7 @@ class ModifiedMatcher extends MatcherHeaderAbstract
     /**
      * Set If-Modified-Since header.
      *
-     * @param string|array $ifModifiedSinceHeaderValue if-modified-since header value or values
+     * @param string|array<string> $ifModifiedSinceHeaderValue if-modified-since header value or values
      */
     public function ifModifiedSinceHeader(string|array $ifModifiedSinceHeaderValue): static
     {
@@ -24,7 +25,7 @@ class ModifiedMatcher extends MatcherHeaderAbstract
     /**
      * New instance with If-Modified-Since header.
      *
-     * @param string|array $ifModifiedSinceHeaderValue if-modified-since header value or values
+     * @param string|array<string> $ifModifiedSinceHeaderValue if-modified-since header value or values
      */
     public function withIfModifiedSinceHeader(string|array $ifModifiedSinceHeaderValue): static
     {
@@ -34,7 +35,7 @@ class ModifiedMatcher extends MatcherHeaderAbstract
     /**
      * Set If-Unmodified-Since header.
      *
-     * @param string|array $ifUnmodifiedSinceHeaderValue if-unmodified-since header value or values
+     * @param string|array<string> $ifUnmodifiedSinceHeaderValue if-unmodified-since header value or values
      */
     public function ifUnmodifiedSinceHeader(string|array $ifUnmodifiedSinceHeaderValue): static
     {
@@ -45,7 +46,7 @@ class ModifiedMatcher extends MatcherHeaderAbstract
     /**
      * New instance with If-Unmodified-Since header.
      *
-     * @param string|array $ifUnmodifiedSinceHeaderValue if-unmodified-since header value or values
+     * @param string|array<string> $ifUnmodifiedSinceHeaderValue if-unmodified-since header value or values
      */
     public function withIfUnmodifiedSinceHeader(string|array $ifUnmodifiedSinceHeaderValue): static
     {
@@ -77,7 +78,7 @@ class ModifiedMatcher extends MatcherHeaderAbstract
     }
 
     /**
-     * @throws \DateMalformedStringException
+     * @throws DateMalformedStringException
      */
     public function getIfModifiedSinceHeaderAsTimestamp(): ?int
     {
@@ -87,7 +88,11 @@ class ModifiedMatcher extends MatcherHeaderAbstract
         if (!$this->isValidIfModifiedSinceHeader()) {
             throw new \RuntimeException('Invalid If-Modified-Since header value');
         }
-        return TimeHelper::toTimestamp($this->getIfModifiedSinceHeader());
+        $time = $this->getIfModifiedSinceHeader();
+        if (null === $time) {
+            throw new \LogicException('If-Modified-Since header is empty');
+        }
+        return TimeHelper::toTimestamp($time);
     }
 
     public function getIfUnmodifiedSinceHeader(): ?string
@@ -101,7 +106,7 @@ class ModifiedMatcher extends MatcherHeaderAbstract
     }
 
     /**
-     * @throws \DateMalformedStringException
+     * @throws DateMalformedStringException
      */
     public function getIfUnmodifiedSinceHeaderAsTimestamp(): ?int
     {
@@ -111,7 +116,11 @@ class ModifiedMatcher extends MatcherHeaderAbstract
         if (!$this->isValidIfUnmodifiedSinceHeader()) {
             throw new \RuntimeException('Invalid If-Unmodified-Since header value');
         }
-        return TimeHelper::toTimestamp($this->getIfUnmodifiedSinceHeader());
+        $time = $this->getIfUnmodifiedSinceHeader();
+        if (null === $time) {
+            throw new \LogicException('If-Unmodified-Since header is empty');
+        }
+        return TimeHelper::toTimestamp($time);
     }
 
     public function isValidIfUnmodifiedSinceHeader(): bool
@@ -130,7 +139,7 @@ class ModifiedMatcher extends MatcherHeaderAbstract
 
     /**
      * Modified match result.
-     * @throws \DateMalformedStringException
+     * @throws DateMalformedStringException
      */
     public function matches(int|string|\DateTime $baseData): ModifiedMatcherResult
     {